fix build order
