+++ scripts/buildPackages/index.js
@@ -1,0 +1,6 @@
+const childProcess = require('child_process');
+
+const path = 'scripts/buildPackages';
+
+childProcess.execSync(`node ${path}/buildComponentsPackages`);
+childProcess.execSync(`node ${path}/buildCustomPackages`);
